- Zmiana metody rejestracji - Zmiana metody hashowania haseł - Wstępny system uprawnień

+++ users/delete.php
@@ -1,21 +1,23 @@
 <?php
-require_once('../config.php');
-global $pdo;
-try {
-    $stmt = $pdo->prepare('DELETE FROM users WHERE ID = :id');
-    $stmt->bindValue(":id", $id, PDO::PARAM_STR);
-    $stmt->execute();
+    require_once('../config.php');
+    global $pdo;
+    session_start();
+    if (isset($_SESSION['logged'])) {
+        try {
+            $stmt = $pdo->prepare('DELETE FROM users WHERE ID = :id');
+            $stmt->bindValue(":id", $id, PDO::PARAM_STR);
+            $stmt->execute();
 
-}
-catch (PDOException $e)
-{
-    echo 'Wystąpił błąd' . $e ->getMessage();
-}
-if($stmt -> rowCount() != 0)
-{
-    echo '<div id="div3" class="alert alert-success" role="alert">Usunięto!</div> ';
-}
-else{
-    echo ' <div id="div3" class="alert alert-danger" role="alert">Wystąpił błąd przy usuwaniu!</div>';
-}
-$stmt -> closeCursor();
+        } catch (PDOException $e) {
+            echo 'Wystąpił błąd' . $e->getMessage();
+        }
+        if ($stmt->rowCount() != 0) {
+            echo '<div id="div3" class="alert alert-success" role="alert">Usunięto!</div> ';
+        } else {
+            echo ' <div id="div3" class="alert alert-danger" role="alert">Wystąpił błąd przy usuwaniu!</div>';
+        }
+        $stmt->closeCursor();
+    }
+    else{
+        header('Location: ../index.php');
+    }
